refactor(native): split runtime controller modules and tests
