+++ src/native/engine/runtime/engine/engine_lifecycle.c
@@ -1,6 +1,6 @@
 #include "engine_lifecycle.h"
 
-#include "../controller/controller_kind_domain.h"
+#include "../controller/kind/controller_kind_domain.h"
 #include "../terrain/terrain_generation.h"
 
 #include <stddef.h>
