+++ src/native/engine/runtime/parallel/parallel_contract.c
@@ -8,8 +8,8 @@ static const RuntimeParallelBudget kRuntimeParallelBudgets[] = {
     {"physics.dynamics.integrate", "src/native/engine/physics/dynamics/dynamics.c", 54, 0.45},
     {"physics.world_collision.sweep", "src/native/engine/physics/world/world_collision.c", 15, 0.40},
     {"world.query.nearby", "src/native/engine/world/query/world_query.c", 26, 0.52},
-    {"runtime.controller_sync.update", "src/native/engine/runtime/controller/controller_sync.c", 55, 0.50},
-    {"runtime.controller_sync.entity_sync", "src/native/engine/runtime/controller/controller_sync.c", 75, 0.45},
+    {"runtime.controller_sync.update", "src/native/engine/runtime/controller/sync/controller_sync.c", 55, 0.50},
+    {"runtime.controller_sync.entity_sync", "src/native/engine/runtime/controller/sync/controller_sync.c", 75, 0.45},
     {"render.mesh_terrain.vertices", "src/native/engine/render/mesh/mesh_terrain.c", 30, 0.55},
     {"render.mesh_terrain.normals", "src/native/engine/render/mesh/mesh_terrain.c", 74, 0.55},
     {"render.renderer.instance_upload", "src/native/engine/render/renderer/renderer.c", 259, 0.64},
