+++ src/native/engine/runtime/player/player_api.c
@@ -1,6 +1,6 @@
 #include "player_api.h"
 #include "player_builds.h"
-#include "../controller/controller_kind_domain.h"
+#include "../controller/kind/controller_kind_domain.h"
 
 static BuildClass default_class_for_controller(const char *controller_kind)
 {
