+++ tests/native/runtime/controller/attach/runtime_controller_attach_test.c
@@ -1,4 +1,4 @@
-#include "runtime/controller/controller_attach.h"
+#include "runtime/controller/attach/controller_attach.h"
 
 #include <stdio.h>
 #include <stdlib.h>
