+++ tests/native/runtime/controller/sync/runtime_controller_sync_strict_test.c
@@ -1,4 +1,4 @@
-#include "runtime/controller/controller_sync.h"
+#include "runtime/controller/sync/controller_sync.h"
 
 #include <stdio.h>
